+++ app/keepers.go
@@ -6,6 +6,7 @@ import (
 
 	wasmdapp "github.com/CosmWasm/wasmd/app"
 	"github.com/CosmWasm/wasmd/x/wasm"
+	wasmkeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
 	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
 	_ "github.com/cosmos/cosmos-sdk/client/docs/statik"
 	"github.com/cosmos/cosmos-sdk/client/flags"
@@ -169,7 +170,7 @@ type AppKeepers struct {
 	TokenFactoryKeeper tokenfactorykeeper.Keeper
 
 	// WASM keepers
-	WasmKeeper       wasm.Keeper
+	WasmKeeper       wasmkeeper.Keeper
 	ScopedWasmKeeper capabilitykeeper.ScopedKeeper
 }
 
@@ -204,7 +205,7 @@ func initStoreKeys() (
 		epochstypes.StoreKey,
 		inflationtypes.StoreKey,
 		sudotypes.StoreKey,
-		wasm.StoreKey,
+		wasmtypes.StoreKey,
 		devgastypes.StoreKey,
 		tokenfactorytypes.StoreKey,
 	)
@@ -382,7 +383,7 @@ func (app *NibiruApp) InitKeepers(
 		app.BankKeeper,
 	)
 
-	app.ScopedWasmKeeper = app.capabilityKeeper.ScopeToModule(wasm.ModuleName)
+	app.ScopedWasmKeeper = app.capabilityKeeper.ScopeToModule(wasmtypes.ModuleName)
 
 	wasmDir := filepath.Join(homePath, "data")
 	wasmConfig, err := wasm.ReadWasmConfig(appOpts)
@@ -399,9 +400,9 @@ func (app *NibiruApp) InitKeepers(
 	// For example, if there are bindings for the x/perp module, then the app
 	// passed to GetWasmOpts must already have a non-nil PerpKeeper.
 	supportedFeatures := strings.Join(wasmdapp.AllCapabilities(), ",")
-	app.WasmKeeper = wasm.NewKeeper(
+	app.WasmKeeper = wasmkeeper.NewKeeper(
 		appCodec,
-		keys[wasm.StoreKey],
+		keys[wasmtypes.StoreKey],
 		app.AccountKeeper,
 		app.BankKeeper,
 		app.stakingKeeper,
@@ -417,7 +418,7 @@ func (app *NibiruApp) InitKeepers(
 		wasmConfig,
 		supportedFeatures,
 		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
-		GetWasmOpts(appOpts, *app)...,
+		GetWasmOpts(*app, appOpts)...,
 	)
 
 	// DevGas uses WasmKeeper
@@ -644,7 +645,7 @@ func orderedModuleNames() []string {
 
 		// --------------------------------------------------------------------
 		// CosmWasm
-		wasm.ModuleName,
+		wasmtypes.ModuleName,
 		devgastypes.ModuleName,
 		tokenfactorytypes.ModuleName,
 
@@ -772,7 +773,7 @@ func ModuleAccPerms() map[string][]string {
 		epochstypes.ModuleName:           {},
 		sudotypes.ModuleName:             {},
 		common.TreasuryPoolModuleAccount: {},
-		wasm.ModuleName:                  {authtypes.Burner},
+		wasmtypes.ModuleName:             {authtypes.Burner},
 		tokenfactorytypes.ModuleName:     {authtypes.Minter, authtypes.Burner},
 	}
 }
@@ -799,7 +800,7 @@ func initParamsKeeper(
 	paramsKeeper.Subspace(ibcexported.ModuleName)
 	paramsKeeper.Subspace(ibcfeetypes.ModuleName)
 	// wasm params keepers
-	paramsKeeper.Subspace(wasm.ModuleName)
+	paramsKeeper.Subspace(wasmtypes.ModuleName)
 	paramsKeeper.Subspace(devgastypes.ModuleName)
 
 	return paramsKeeper